requirements.txt
Flask==2.2.5
SQLAlchemy==1.4.52
pyodbc==4.0.34
python-dotenv==1.0.0
Flask-WTF==1.1.1
WTForms==3.0.1